feat: 키보드 상태 함수 분리
