+++ app/src/main/java/com/echoist/linkedout/page/myLog/WritingPage.kt
@@ -87,6 +87,7 @@ import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
+import com.echoist.linkedout.Keyboard
 import com.echoist.linkedout.R
 import com.echoist.linkedout.Routes
 import com.echoist.linkedout.api.EssayApi
@@ -102,8 +103,7 @@ import com.echoist.linkedout.components.LocationGroup
 import com.echoist.linkedout.components.LocationTextField
 import com.echoist.linkedout.components.TextItem
 import com.echoist.linkedout.gps.RequestPermissionsUtil
-import com.echoist.linkedout.page.login.Keyboard
-import com.echoist.linkedout.page.login.keyboardAsState
+import com.echoist.linkedout.keyboardAsState
 import com.echoist.linkedout.ui.theme.LinkedInColor
 import com.echoist.linkedout.viewModels.WritingViewModel
 import com.mohamedrejeb.richeditor.model.RichTextState
